+++ src/corelib/packet/statspacket.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2014-2016 Fanout, Inc.
+ * Copyright (C) 2014-2017 Fanout, Inc.
  *
  * This file is part of Pushpin.
  *
@@ -86,6 +86,9 @@ QVariant StatsPacket::toVariant() const
 		if(type == Subscribed)
 		{
 			obj["ttl"] = ttl;
+
+			if(subscribers >= 0)
+				obj["subscribers"] = subscribers;
 		}
 		else // Unsubscribed
 		{
@@ -108,6 +111,8 @@ QVariant StatsPacket::toVariant() const
 			obj["blocks-received"] = blocksReceived;
 		if(blocksSent >= 0)
 			obj["blocks-sent"] = blocksSent;
+		if(duration >= 0)
+			obj["duration"] = duration;
 	}
 
 	return obj;
@@ -270,6 +275,16 @@ bool StatsPacket::fromVariant(const QByteArray &_type, const QVariant &in)
 			ttl = obj["ttl"].toInt();
 			if(ttl < 0)
 				return false;
+
+			if(obj.contains("subscribers"))
+			{
+				if(!obj["subscribers"].canConvert(QVariant::Int))
+					return false;
+
+				subscribers = obj["subscribers"].toInt();
+				if(subscribers < 0)
+					return false;
+			}
 		}
 	}
 	else if(_type == "report")
@@ -328,6 +343,14 @@ bool StatsPacket::fromVariant(const QByteArray &_type, const QVariant &in)
 				return false;
 
 			blocksSent = obj["blocks-sent"].toInt();
+		}
+
+		if(obj.contains("duration"))
+		{
+			if(!obj["duration"].canConvert(QVariant::Int))
+				return false;
+
+			duration = obj["duration"].toInt();
 		}
 	}
 	else