+++ src/corelib/statsmanager.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2014-2016 Fanout, Inc.
+ * Copyright (C) 2014-2017 Fanout, Inc.
  *
  * This file is part of Pushpin.
  *
@@ -20,6 +20,7 @@
 #include "statsmanager.h"
 
 #include <assert.h>
+#include <QVector>
 #include <QDateTime>
 #include <QPointer>
 #include <QTimer>
@@ -33,21 +34,10 @@
 #define OUT_HWM 200000
 
 #define ACTIVITY_TIMEOUT 100
-#define REPORT_INTERVAL (10 * 1000)
-
 #define REFRESH_INTERVAL 1000
-#define CONNECTION_EXPIRE (120 * 1000)
-#define CONNECTION_LINGER (60 * 1000)
-#define SUBSCRIPTION_EXPIRE (60 * 1000)
-#define SUBSCRIPTION_LINGER (60 * 1000)
 
-#define CONNECTION_SHOULD_PROCESS (CONNECTION_EXPIRE * 3 / 4)
-#define CONNECTION_MUST_PROCESS (CONNECTION_EXPIRE * 4 / 5)
-#define CONNECTION_REFRESH_BUCKETS (CONNECTION_SHOULD_PROCESS / REFRESH_INTERVAL)
-
-#define SUBSCRIPTION_SHOULD_PROCESS (SUBSCRIPTION_EXPIRE * 3 / 4)
-#define SUBSCRIPTION_MUST_PROCESS (SUBSCRIPTION_EXPIRE * 4 / 5)
-#define SUBSCRIPTION_REFRESH_BUCKETS (SUBSCRIPTION_SHOULD_PROCESS / REFRESH_INTERVAL)
+#define SHOULD_PROCESS_TIME(x) (x * 3 / 4)
+#define MUST_PROCESS_TIME(x) (x * 4 / 5)
 
 class StatsManager::Private : public QObject
 {
@@ -87,11 +77,13 @@ public:
 	public:
 		QString mode;
 		QString channel;
+		int subscriberCount;
 		qint64 lastRefresh;
 		int refreshBucket;
 		bool linger;
 
 		Subscription() :
+			subscriberCount(0),
 			lastRefresh(-1),
 			refreshBucket(-1),
 			linger(false)
@@ -112,6 +104,7 @@ public:
 		int blocksReceived;
 		int blocksSent;
 		qint64 lastUpdate;
+		qint64 startTime;
 
 		Report() :
 			connectionsMax(0),
@@ -122,7 +115,8 @@ public:
 			httpResponseMessagesSent(0),
 			blocksReceived(-1),
 			blocksSent(-1),
-			lastUpdate(-1)
+			lastUpdate(-1),
+			startTime(-1)
 		{
 		}
 
@@ -144,35 +138,43 @@ public:
 	QByteArray instanceId;
 	int ipcFileMode;
 	QString spec;
+	int connectionTtl;
+	int connectionLinger;
+	int subscriptionTtl;
+	int subscriptionLinger;
+	int reportInterval;
 	QZmq::Socket *sock;
 	QHash<QByteArray, int> routeActivity;
 	QHash<QByteArray, ConnectionInfo*> connectionInfoById;
 	QHash<QByteArray, QSet<ConnectionInfo*> > connectionInfoByRoute;
 	QMap<QPair<qint64, ConnectionInfo*>, ConnectionInfo*> connectionInfoByLastRefresh;
-	QSet<ConnectionInfo*> connectionInfoRefreshBuckets[CONNECTION_REFRESH_BUCKETS];
+	QVector<QSet<ConnectionInfo*> > connectionInfoRefreshBuckets;
 	int currentConnectionInfoRefreshBucket;
 	QHash<QByteArray, QHash<QByteArray, ConnectionInfo*> > externalConnectionInfoByFrom;
 	QHash<QByteArray, QSet<ConnectionInfo*> > externalConnectionInfoByRoute;
 	QMap<QPair<qint64, ConnectionInfo*>, ConnectionInfo*> externalConnectionInfoByLastActive;
 	QHash<SubscriptionKey, Subscription*> subscriptionsByKey;
 	QMap<QPair<qint64, Subscription*>, Subscription*> subscriptionsByLastRefresh;
-	QSet<Subscription*> subscriptionRefreshBuckets[SUBSCRIPTION_REFRESH_BUCKETS];
+	QVector<QSet<Subscription*> > subscriptionRefreshBuckets;
 	int currentSubscriptionRefreshBucket;
 	QHash<QByteArray, Report*> reports;
 	QTimer *activityTimer;
 	QTimer *reportTimer;
 	QTimer *refreshTimer;
-	bool reportsEnabled;
 
 	Private(StatsManager *_q) :
 		QObject(_q),
 		q(_q),
 		ipcFileMode(-1),
+		connectionTtl(120 * 1000),
+		connectionLinger(60 * 1000),
+		subscriptionTtl(60 * 1000),
+		subscriptionLinger(60 * 1000),
+		reportInterval(10 * 1000),
 		sock(0),
 		currentConnectionInfoRefreshBucket(0),
 		currentSubscriptionRefreshBucket(0),
-		reportTimer(0),
-		reportsEnabled(false)
+		reportTimer(0)
 	{
 		activityTimer = new QTimer(this);
 		connect(activityTimer, &QTimer::timeout, this, &Private::activity_timeout);
@@ -181,6 +183,9 @@ public:
 		refreshTimer = new QTimer(this);
 		connect(refreshTimer, &QTimer::timeout, this, &Private::refresh_timeout);
 		refreshTimer->start(REFRESH_INTERVAL);
+
+		setupConnectionBuckets();
+		setupSubscriptionBuckets();
 	}
 
 	~Private()
@@ -223,7 +228,7 @@ public:
 		qDeleteAll(reports);
 	}
 
-	bool setup()
+	bool setupSock()
 	{
 		delete sock;
 
@@ -243,15 +248,57 @@ public:
 		return true;
 	}
 
-	void setupReports()
+	void setupConnectionBuckets()
 	{
-		if(reportsEnabled && !reportTimer)
+		int shouldProcessTime = SHOULD_PROCESS_TIME(connectionTtl);
+		QVector<QSet<ConnectionInfo*> > newBuckets(qMax(shouldProcessTime / REFRESH_INTERVAL, 1));
+
+		// rebalance (NOTE: this algorithm is not optimal)
+		int nextBucketIndex = 0;
+		for(int n = 0; n < connectionInfoRefreshBuckets.count(); ++n)
+		{
+			foreach(ConnectionInfo *c, connectionInfoRefreshBuckets[n])
+			{
+				newBuckets[nextBucketIndex++] += c;
+				if(nextBucketIndex >= newBuckets.count())
+					nextBucketIndex = 0;
+			}
+		}
+
+		connectionInfoRefreshBuckets = newBuckets;
+		currentConnectionInfoRefreshBucket = 0;
+	}
+
+	void setupSubscriptionBuckets()
+	{
+		int shouldProcessTime = SHOULD_PROCESS_TIME(subscriptionTtl);
+		QVector<QSet<Subscription*> > newBuckets(qMax(shouldProcessTime / REFRESH_INTERVAL, 1));
+
+		// rebalance (NOTE: this algorithm is not optimal)
+		int nextBucketIndex = 0;
+		for(int n = 0; n < subscriptionRefreshBuckets.count(); ++n)
+		{
+			foreach(Subscription *s, subscriptionRefreshBuckets[n])
+			{
+				newBuckets[nextBucketIndex++] += s;
+				if(nextBucketIndex >= newBuckets.count())
+					nextBucketIndex = 0;
+			}
+		}
+
+		subscriptionRefreshBuckets = newBuckets;
+		currentSubscriptionRefreshBucket = 0;
+	}
+
+	void setupReportTimer()
+	{
+		if(reportInterval > 0 && !reportTimer)
 		{
 			reportTimer = new QTimer(this);
 			connect(reportTimer, &QTimer::timeout, this, &Private::report_timeout);
-			reportTimer->start(REPORT_INTERVAL);
+			reportTimer->start(reportInterval);
 		}
-		else if(!reportsEnabled && reportTimer)
+		else if(reportInterval <= 0 && reportTimer)
 		{
 			reportTimer->setParent(0);
 			reportTimer->disconnect(this);
@@ -264,7 +311,7 @@ public:
 	{
 		int best = -1;
 		int bestSize;
-		for(int n = 0; n < CONNECTION_REFRESH_BUCKETS; ++n)
+		for(int n = 0; n < connectionInfoRefreshBuckets.count(); ++n)
 		{
 			if(best == -1 || connectionInfoRefreshBuckets[n].count() < bestSize)
 			{
@@ -280,7 +327,7 @@ public:
 	{
 		int best = -1;
 		int bestSize;
-		for(int n = 0; n < SUBSCRIPTION_REFRESH_BUCKETS; ++n)
+		for(int n = 0; n < subscriptionRefreshBuckets.count(); ++n)
 		{
 			if(best == -1 || subscriptionRefreshBuckets[n].count() < bestSize)
 			{
@@ -397,6 +444,7 @@ public:
 		{
 			report = new Report;
 			report->routeId = routeId;
+			report->startTime = QDateTime::currentMSecsSinceEpoch();
 			reports[routeId] = report;
 		}
 
@@ -478,7 +526,7 @@ public:
 			p.connectionType = StatsPacket::Http;
 		p.peerAddress = c->peerAddress;
 		p.ssl = c->ssl;
-		p.ttl = CONNECTION_EXPIRE / 1000;
+		p.ttl = connectionTtl / 1000;
 		write(p);
 	}
 
@@ -505,7 +553,8 @@ public:
 		p.from = instanceId;
 		p.mode = s->mode.toUtf8();
 		p.channel = s->channel.toUtf8();
-		p.ttl = SUBSCRIPTION_EXPIRE / 1000;
+		p.ttl = subscriptionTtl / 1000;
+		p.subscribers = s->subscriberCount;
 		write(p);
 	}
 
@@ -585,7 +634,7 @@ public:
 		}
 
 		// process any others
-		qint64 threshold = now - CONNECTION_MUST_PROCESS;
+		qint64 threshold = now - MUST_PROCESS_TIME(connectionTtl);
 		while(!connectionInfoByLastRefresh.isEmpty())
 		{
 			QMap<QPair<qint64, ConnectionInfo*>, ConnectionInfo*>::iterator it = connectionInfoByLastRefresh.begin();
@@ -639,19 +688,19 @@ public:
 			emit q->connectionsRefreshed(refreshedIds);
 
 		++currentConnectionInfoRefreshBucket;
-		if(currentConnectionInfoRefreshBucket >= CONNECTION_REFRESH_BUCKETS)
+		if(currentConnectionInfoRefreshBucket >= connectionInfoRefreshBuckets.count())
 			currentConnectionInfoRefreshBucket = 0;
 	}
 
 	void expireExternalConnections(qint64 now)
 	{
 		// external connections should only be tracked if reporting is enabled
-		if(!reportsEnabled)
+		if(reportInterval <= 0)
 			return;
 
 		QSet<QByteArray> routesUpdated;
 
-		qint64 threshold = now - CONNECTION_EXPIRE;
+		qint64 threshold = now - connectionTtl;
 		while(!externalConnectionInfoByLastActive.isEmpty())
 		{
 			QMap<QPair<qint64, ConnectionInfo*>, ConnectionInfo*>::iterator it = externalConnectionInfoByLastActive.begin();
@@ -693,7 +742,7 @@ public:
 		}
 
 		// process any others
-		qint64 threshold = now - SUBSCRIPTION_MUST_PROCESS;
+		qint64 threshold = now - MUST_PROCESS_TIME(subscriptionTtl);
 		while(!subscriptionsByLastRefresh.isEmpty())
 		{
 			QMap<QPair<qint64, Subscription*>, Subscription*>::iterator it = subscriptionsByLastRefresh.begin();
@@ -740,7 +789,7 @@ public:
 		}
 
 		++currentSubscriptionRefreshBucket;
-		if(currentSubscriptionRefreshBucket >= SUBSCRIPTION_REFRESH_BUCKETS)
+		if(currentSubscriptionRefreshBucket >= subscriptionRefreshBuckets.count())
 			currentSubscriptionRefreshBucket = 0;
 	}
 
@@ -789,7 +838,9 @@ private slots:
 			p.httpResponseMessagesSent = report->httpResponseMessagesSent;
 			p.blocksReceived = report->blocksReceived;
 			p.blocksSent = report->blocksSent;
+			p.duration = now - report->startTime;
 
+			report->startTime = now;
 			report->connectionsMaxStale = true;
 			report->connectionsMinutes = 0;
 			report->messagesReceived = 0;
@@ -854,13 +905,30 @@ void StatsManager::setIpcFileMode(int mode)
 bool StatsManager::setSpec(const QString &spec)
 {
 	d->spec = spec;
-	return d->setup();
+	return d->setupSock();
 }
 
-void StatsManager::setReportsEnabled(bool on)
+void StatsManager::setConnectionTtl(int secs)
 {
-	d->reportsEnabled = on;
-	d->setupReports();
+	d->connectionTtl = secs * 1000;
+	d->setupConnectionBuckets();
+}
+
+void StatsManager::setSubscriptionTtl(int secs)
+{
+	d->subscriptionTtl = secs * 1000;
+	d->setupSubscriptionBuckets();
+}
+
+void StatsManager::setSubscriptionLinger(int secs)
+{
+	d->subscriptionLinger = secs * 1000;
+}
+
+void StatsManager::setReportInterval(int secs)
+{
+	d->reportInterval = secs * 1000;
+	d->setupReportTimer();
 }
 
 void StatsManager::addActivity(const QByteArray &routeId, int count)
@@ -906,7 +974,7 @@ void StatsManager::addConnection(const QByteArray &id, const QByteArray &routeId
 	c->lastReport = now; // start counting minutes from current time
 	d->insertConnection(c);
 
-	if(d->reportsEnabled)
+	if(d->reportInterval > 0)
 	{
 		// check if this connection should replace a lingering external one
 		// note: this iterates over all the known external sources, which at
@@ -952,7 +1020,7 @@ void StatsManager::removeConnection(const QByteArray &id, bool linger)
 	qint64 now = QDateTime::currentMSecsSinceEpoch();
 	QByteArray routeId = c->routeId;
 
-	if(d->reportsEnabled)
+	if(d->reportInterval > 0)
 		d->updateConnectionsMinutes(c, now);
 
 	if(linger)
@@ -962,7 +1030,7 @@ void StatsManager::removeConnection(const QByteArray &id, bool linger)
 			c->linger = true;
 
 			// hack to ensure full linger time honored by refresh processing
-			qint64 lingerStartTime = now + (CONNECTION_EXPIRE - CONNECTION_MUST_PROCESS);
+			qint64 lingerStartTime = now + (d->connectionLinger - MUST_PROCESS_TIME(d->connectionTtl));
 
 			// move to the end
 			d->connectionInfoByLastRefresh.remove(QPair<qint64, Private::ConnectionInfo*>(c->lastRefresh, c));
@@ -977,7 +1045,7 @@ void StatsManager::removeConnection(const QByteArray &id, bool linger)
 		delete c;
 	}
 
-	if(d->reportsEnabled)
+	if(d->reportInterval > 0)
 		d->updateConnectionsMax(routeId, now);
 }
 
@@ -990,7 +1058,7 @@ void StatsManager::refreshConnection(const QByteArray &id)
 	d->sendConnected(c);
 }
 
-void StatsManager::addSubscription(const QString &mode, const QString &channel)
+void StatsManager::addSubscription(const QString &mode, const QString &channel, int subscriberCount)
 {
 	Private::SubscriptionKey subKey(mode, channel);
 	Private::Subscription *s = d->subscriptionsByKey.value(subKey);
@@ -1002,6 +1070,7 @@ void StatsManager::addSubscription(const QString &mode, const QString &channel)
 		s = new Private::Subscription;
 		s->mode = mode;
 		s->channel = channel;
+		s->subscriberCount = subscriberCount;
 		s->lastRefresh = now;
 		d->insertSubscription(s);
 
@@ -1009,6 +1078,10 @@ void StatsManager::addSubscription(const QString &mode, const QString &channel)
 	}
 	else
 	{
+		int oldSubscriberCount = s->subscriberCount;
+
+		s->subscriberCount = subscriberCount;
+
 		if(s->linger)
 		{
 			qint64 now = QDateTime::currentMSecsSinceEpoch();
@@ -1022,6 +1095,15 @@ void StatsManager::addSubscription(const QString &mode, const QString &channel)
 			d->subscriptionsByLastRefresh.insert(QPair<qint64, Private::Subscription*>(s->lastRefresh, s), s);
 
 			d->sendSubscribed(s);
+		}
+		else if(s->subscriberCount != oldSubscriberCount)
+		{
+			qint64 now = QDateTime::currentMSecsSinceEpoch();
+
+			// process soon
+			d->subscriptionsByLastRefresh.remove(QPair<qint64, Private::Subscription*>(s->lastRefresh, s));
+			s->lastRefresh = now - MUST_PROCESS_TIME(d->subscriptionTtl);
+			d->subscriptionsByLastRefresh.insert(QPair<qint64, Private::Subscription*>(s->lastRefresh, s), s);
 		}
 	}
 }
@@ -1042,7 +1124,7 @@ void StatsManager::removeSubscription(const QString &mode, const QString &channe
 			s->linger = true;
 
 			// hack to ensure full linger time honored by refresh processing
-			qint64 lingerStartTime = now + (SUBSCRIPTION_EXPIRE - SUBSCRIPTION_MUST_PROCESS);
+			qint64 lingerStartTime = now + (d->subscriptionLinger - MUST_PROCESS_TIME(d->subscriptionTtl));
 
 			// move to the end
 			d->subscriptionsByLastRefresh.remove(QPair<qint64, Private::Subscription*>(s->lastRefresh, s));
@@ -1062,7 +1144,8 @@ void StatsManager::removeSubscription(const QString &mode, const QString &channe
 
 void StatsManager::addMessageReceived(const QByteArray &routeId, int blocks)
 {
-	assert(d->reportsEnabled);
+	if(d->reportInterval <= 0)
+		return;
 
 	Private::Report *report = d->getOrCreateReport(routeId);
 
@@ -1081,7 +1164,8 @@ void StatsManager::addMessageReceived(const QByteArray &routeId, int blocks)
 
 void StatsManager::addMessageSent(const QByteArray &routeId, const QString &transport, int blocks)
 {
-	assert(d->reportsEnabled);
+	if(d->reportInterval <= 0)
+		return;
 
 	Private::Report *report = d->getOrCreateReport(routeId);
 
@@ -1108,7 +1192,8 @@ bool StatsManager::checkConnection(const QByteArray &id)
 
 void StatsManager::processExternalPacket(const StatsPacket &packet)
 {
-	assert(d->reportsEnabled);
+	if(d->reportInterval <= 0)
+		return;
 
 	if(packet.type != StatsPacket::Connected && packet.type != StatsPacket::Disconnected)
 		return;