+++ src/proxy/engine.h
@@ -71,6 +71,8 @@ public:
 		QString sockJsUrl;
 		QString updatesCheck;
 		QString organizationName;
+		bool quietCheck;
+		int statsConnectionTtl;
 
 		Configuration() :
 			ipcFileMode(-1),
@@ -83,7 +85,9 @@ public:
 			useXForwardedProtocol(false),
 			logFrom(false),
 			logUserAgent(false),
-			updatesCheck("check")
+			updatesCheck("check"),
+			quietCheck(false),
+			statsConnectionTtl(-1)
 		{
 		}
 	};